+++ packages/undead/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.2-wip
+
+- Add `ensure_cli_readme_test.dart` to verify CLI `--help` documentation in `README.md`.
+
 ## 0.1.1
 
 - Support `--suggest-private` flag to detect unexported top-level declarations